backend/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# File Processing
PyMuPDF==1.23.14
ebooklib==0.18

# RAG & LLM
langchain==0.0.352
chromadb==0.4.22
sentence-transformers==2.2.2
google-generativeai==0.3.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
typing-extensions==4.8.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2 